move grid module to components/
... so that its attributes are added to the schema
    free of circular deps.
