+++ src/plots/layout_attributes.js
@@ -10,7 +10,6 @@
 
 var fontAttrs = require('./font_attributes');
 var colorAttrs = require('../components/color/attributes');
-var gridAttrs = require('./grid').attributes;
 
 var globalFont = fontAttrs({
     editType: 'calc',
@@ -196,6 +195,5 @@ module.exports = {
             'being treated as immutable, thus any data array with a',
             'different identity from its predecessor contains new data.'
         ].join(' ')
-    },
-    grid: gridAttrs
+    }
 };
